requirements.txt
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
Django==2.1.15
django-bootstrap-modal-forms==2.0.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-extensions==3.0.3
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-widget-tweaks==1.4.8
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
